Assignment abut Zoo. practices GUI
Animal in Zoo. can eat food (and can eat one another, if one is a carnivore). you can move the animal, watch at the Zoo animals' info, remove all the animals from the zoo, and exit the system. (PDF document of the assignment in Hebrew , you can add a comment if there will be need a translation)